+++ include/cedar/button.h
@@ -1,6 +1,10 @@
 #ifndef _LIBCEDAR_BUTTON_H_
 #define _LIBCEDAR_BUTTON_H_
 
+#ifdef __cplusplus
+extern "C" {
+#endif
+
 #include <cedar.h>
 
 // TODO: either prefix with Cedar or dont expose to user
@@ -18,5 +22,9 @@ typedef struct {
  * height	The height of the button
  */
 CedarWidget *CedarButton(ID id, int x, int y, int width, int height, const char *text);
+
+#ifdef __cplusplus
+}
+#endif
 
 #endif  // _LIBCEDAR_BUTTON_H_
