+++ include/cedar/utils.h
@@ -1,7 +1,11 @@
 #ifndef _CEDAR_UTILS_H_
 #define _CEDAR_UTILS_H_
 
-#include "cedar.h"
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#include <cedar.h>
 
 /*
  * Draw text wrapped to fit in a box
@@ -13,5 +17,9 @@
  * height	The height of the box
  */
 void cedar_wrapTextInBox(const char *str, int x, int y, unsigned int width, unsigned int height);
+
+#ifdef __cplusplus
+}
+#endif
 
 #endif  // _CEDAR_UTILS_H_
